fix : QClass rename fix
