+++ src/main/java/bitcopark/library/repository/book/BookRepositoryCustomImpl.java
@@ -2,6 +2,7 @@ package bitcopark.library.repository.book;
 
 import bitcopark.library.dto.BookSearchDetailCondition;
 import bitcopark.library.dto.BookSearchDto;
+import bitcopark.library.dto.QBookSearchDto;
 import bitcopark.library.entity.book.QBook;
 import bitcopark.library.entity.book.QBookBorrow;
 import bitcopark.library.entity.book.QBookReservation;
